requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
crispy-bootstrap5==2024.2
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-bootstrap4==23.2
django-countries==7.5.1
django-crispy-forms==2.0
django-utils-six==2.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.0
oauthlib==3.2.2
orjson==3.10.6
packaging==23.2
pandas==2.1.1
pip-review==1.3.0
plotly==5.17.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
starlette==0.37.2
tenacity==8.2.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.0.6
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.2
